Backend/requirements.txt
fastapi
uvicorn
python-dotenv
huggingface-hub
pillow
python-multipart
